From [#985544] in comment #12.
We can either revert to XML to deal with PHP choking on json_decode() with bigints, or we can just preg_replace the json before we decode it.

There's a nice patch which restores the old XML support in issue #985544, but this workaround is much smaller - though it's a bit of a special case approach.